+++ src/ECommerce.Infrastructure/DependencyInjection.cs
@@ -104,12 +104,15 @@ public static class DependencyInjection
         services.AddScoped<IRoleRepository, RoleRepository>();
         services.AddScoped<IAddressRepository, AddressRepository>();
         services.AddScoped<IRefreshTokenRepository, RefreshTokenRepository>();
+        services.AddScoped<IMfaService, TotpMfaService>();
+        services.AddScoped<IMfaSecretRepository, MfaSecretRepository>();
         services.AddScoped<IProductRepository, ProductRepository>();
         services.AddScoped<ICategoryRepository, CategoryRepository>();
         services.AddScoped<IBrandRepository, BrandRepository>();
         services.AddScoped<IProductImportRepository, ProductImportRepository>();
         services.AddScoped<IProductImportJobScheduler, HangfireProductImportJobScheduler>();
         services.AddScoped<IProductSearchRepository, ProductSearchRepository>();
+        services.AddScoped<IAutocompleteRepository, AutocompleteRepository>();
         services.AddScoped<IEventHandler<ProductCreated>, ProductSearchIndexSynchronizer>();
         services.AddScoped<IEventHandler<ProductUpdated>, ProductSearchIndexSynchronizer>();
         services.AddScoped<IEventHandler<ProductDeactivated>, ProductSearchIndexSynchronizer>();