+++ src/ECommerce.UseCases/Common/ICorrelationIdProvider.cs
@@ -1,0 +1,6 @@
+namespace ECommerce.UseCases.Common;
+
+public interface ICorrelationIdProvider
+{
+    string CorrelationId { get; }
+}