+++ src/ECommerce.UseCases/Identity/Ports/IMfaSecretRepository.cs
@@ -1,0 +1,9 @@
+using ECommerce.Domain.Identity;
+
+namespace ECommerce.UseCases.Identity.Ports;
+
+public interface IMfaSecretRepository
+{
+    Task<MfaSecret?> GetByCustomerIdAsync(Guid customerId, CancellationToken cancellationToken);
+    void Add(MfaSecret mfaSecret);
+}